refactor: 메인 로직의 분기 처리를 enum을 활용한 방식으로 변경
