--- src/main/java/domain/command/Command.java
@@ -1,6 +1,0 @@
-package domain.command;
-
-public interface Command {
-
-    void execute();
-}
